+++ examples/c/select.c
@@ -3,6 +3,7 @@
  * http://creativecommons.org/publicdomain/zero/1.0/ */
 #include <stdio.h>
 #include <stdlib.h>
+#include <inttypes.h>
 
 #include "chemfiles.h"
 
@@ -13,17 +14,17 @@ int main() {
 
     // Create a selection for all atoms with "Zn" name
     CHFL_SELECTION* selection = chfl_selection("name Zn");
-    size_t matching = 0;
+    uint64_t matching = 0;
 
     // Get the number of matching atoms from the frame
     chfl_selection_evalutate(selection, frame, &matching);
-    printf("We have %zu zinc in the frame\n", matching);
+    printf("We have %"PRIu64" zinc in the frame\n", matching);
     chfl_match_t* matches = malloc(matching * sizeof(chfl_match_t));
 
     // Get the matching atoms
     chfl_selection_matches(selection, matches, matching);
-    for (size_t i=0; i<matching; i++) {
-        printf("%zu is a zinc\n", matches[i].atoms[0]);
+    for (uint64_t i=0; i<matching; i++) {
+        printf("%"PRIu64" is a zinc\n", matches[i].atoms[0]);
     }
 
     chfl_selection_free(selection);
@@ -32,14 +33,14 @@ int main() {
     // Create a selection for multiple atoms
     selection = chfl_selection("angles: name($1) H and name($2) O and name($3) H");
     chfl_selection_evalutate(selection, frame, &matching);
-    printf("We have %zu water in the frame\n", matching);
+    printf("We have %"PRIu64" water in the frame\n", matching);
     matches = malloc(matching * sizeof(chfl_match_t));
 
     // Get the matching atoms
     chfl_selection_matches(selection, matches, matching);
-    for (size_t i=0; i<matching; i++) {
+    for (uint64_t i=0; i<matching; i++) {
         printf(
-            "%zu - %zu - %zu is a water\n",
+            "%"PRIu64" - %"PRIu64" - %"PRIu64" is a water\n",
             matches[i].atoms[0], matches[i].atoms[1], matches[i].atoms[2]
         );
     }
